오타 및 테이블 수정

+++ sql/Enrollments.sql
@@ -2,7 +2,7 @@ CREATE TABLE Enrollments (
     enrollment_id bigint AUTO_INCREMENT PRIMARY KEY,
     user_id bigint NOT NULL,
     subject_id bigint NOT NULL,
-    progress bigint DEFAULT 0,
+    progress float DEFAULT 0,
     is_completed BOOLEAN DEFAULT FALSE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES Users(user_id),
@@ -12,7 +12,7 @@ CREATE TABLE Enrollments (
 
 -- enrollment_id: 고유한 등록 식별자.
 -- user_id: 등록한 사용자의 ID.
--- course_id: 등록한 강좌의 ID.
+-- subject_id: 등록한 강좌의 ID.
 -- progress: 강좌 진행률.
 -- is_completed: 강좌 완료 여부.
 -- created_at: 등록 시각.
